Fixed Buffer bugs & added Slack API/Webhook default options
