+++ copyright.js
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) moroo.
+ * Copyright (c) moroo
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
